Update tests and models to replace PartieSurowca with PartieApollo
- Refactored test files to utilize PartieApollo in place of PartieSurowca, ensuring consistency across the test suite.
- Adjusted model imports and relationships in test_apollo_service.py and test_operations_routes.py to reflect the new PartieApollo structure.
- Removed obsolete references to PartieSurowca in various test cases, enhancing clarity and maintainability.
