+++ alembic/versions/06d7543e74e4_typ_surowca_null_true.py
@@ -1,0 +1,34 @@
+"""typ_surowca null true
+
+Revision ID: 06d7543e74e4
+Revises: 96736366f070
+Create Date: 2025-10-17 19:56:40.086042
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mysql
+
+# revision identifiers, used by Alembic.
+revision: str = '06d7543e74e4'
+down_revision: Union[str, None] = '96736366f070'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('partie_apollo', 'typ_surowca',
+               existing_type=mysql.VARCHAR(length=50),
+               nullable=True)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('partie_apollo', 'typ_surowca',
+               existing_type=mysql.VARCHAR(length=50),
+               nullable=False)
+    # ### end Alembic commands ###
