+++ alembic/versions/a1407ae1987b_operacje_log_id_partii_surowca_fk_2.py
@@ -1,0 +1,30 @@
+"""operacje-log id_partii_surowca FK 2
+
+Revision ID: a1407ae1987b
+Revises: a9d757c041d7
+Create Date: 2025-10-17 20:38:16.588565
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'a1407ae1987b'
+down_revision: Union[str, None] = 'a9d757c041d7'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
