+++ alembic/versions/a9d757c041d7_operacje_log_id_partii_surowca_fk.py
@@ -1,0 +1,30 @@
+"""operacje-log id_partii_surowca FK
+
+Revision ID: a9d757c041d7
+Revises: ff4eb7860ecf
+Create Date: 2025-10-17 20:33:27.149973
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'a9d757c041d7'
+down_revision: Union[str, None] = 'ff4eb7860ecf'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_foreign_key('operacje_log_ibfk_1', 'operacje_log', 'partie_apollo', ['id_partii_surowca'], ['id'], ondelete='SET NULL')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('operacje_log_ibfk_1', 'operacje_log', type_='foreignkey')
+    # ### end Alembic commands ###
