+++ alembic/versions/ff4eb7860ecf_typ_transformacji_null_true.py
@@ -1,0 +1,34 @@
+"""typ_transformacji null true
+
+Revision ID: ff4eb7860ecf
+Revises: 06d7543e74e4
+Create Date: 2025-10-17 20:00:03.292892
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mysql
+
+# revision identifiers, used by Alembic.
+revision: str = 'ff4eb7860ecf'
+down_revision: Union[str, None] = '06d7543e74e4'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('partie_apollo', 'typ_transformacji',
+               existing_type=mysql.ENUM('NOWA', 'TRANSFER', 'KOREKTA', 'ZAKONCZENIE'),
+               nullable=True)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('partie_apollo', 'typ_transformacji',
+               existing_type=mysql.ENUM('NOWA', 'TRANSFER', 'KOREKTA', 'ZAKONCZENIE'),
+               nullable=False)
+    # ### end Alembic commands ###
